Support Original Weights in XY of XYZ Plot
Adds X weights to Y weights. If both X and Y have a value in the same
block then Y's is set to 0; both values are used due to both XY and
YX being generated, but the diagonal shows only the X values in that
case. When X and Y have no overlap then XY is generated and YX just
copies the already generated image into the output.

Implemented duplicate base checker that skips generating an image
that's identical to one already generated. If a previous iteration
already generated a base that's identical to the current one then the
existing image is copied and applied to the output.